+++ easy-retrofit-extension-retry-core/src/main/java/io/github/easyretrofit/extension/retry/core/annotation/EnableRetry.java
@@ -2,9 +2,8 @@ package io.github.easyretrofit.extension.retry.core.annotation;
 
 import io.github.easyretrofit.core.annotation.RetrofitInterceptor;
 import io.github.easyretrofit.core.annotation.RetrofitInterceptorParam;
-import io.github.easyretrofit.core.proxy.BaseFallBack;
+import io.github.easyretrofit.core.delegate.BaseFallBack;
 import io.github.easyretrofit.extension.retry.core.interceptor.RetryInterceptor;
-import io.github.easyretrofit.extension.retry.core.resource.BaseRetryConfig;
 
 import java.lang.annotation.*;
