+++ easy-retrofit-extension-retry-core/src/main/java/io/github/easyretrofit/extension/retry/core/interceptor/RetryExceptionFallBackHandler.java
@@ -3,8 +3,8 @@ package io.github.easyretrofit.extension.retry.core.interceptor;
 import io.github.easyretrofit.core.CDIBeanManager;
 import io.github.easyretrofit.core.exception.RetrofitExtensionException;
 import io.github.easyretrofit.core.extension.InterceptorUtils;
-import io.github.easyretrofit.core.proxy.BaseExceptionDelegate;
-import io.github.easyretrofit.core.proxy.BaseFallBack;
+import io.github.easyretrofit.core.delegate.BaseExceptionDelegate;
+import io.github.easyretrofit.core.delegate.BaseFallBack;
 import io.github.easyretrofit.core.resource.RetrofitApiInterfaceBean;
 import io.github.easyretrofit.extension.retry.core.RetrofitRetryResourceContext;
 import io.github.easyretrofit.extension.retry.core.RetryException;
