+++ easy-retrofit-extension-retry-spring-boot-starter/src/main/java/io/github/easyretrofit/extension/retry/spring/boot/RetrofitRetryExtension.java
@@ -3,7 +3,7 @@ package io.github.easyretrofit.extension.retry.spring.boot;
 import io.github.easyretrofit.core.RetrofitInterceptorExtension;
 import io.github.easyretrofit.core.exception.RetrofitExtensionException;
 import io.github.easyretrofit.core.extension.BaseInterceptor;
-import io.github.easyretrofit.core.proxy.BaseExceptionDelegate;
+import io.github.easyretrofit.core.delegate.BaseExceptionDelegate;
 import io.github.easyretrofit.extension.retry.core.annotation.EnableRetry;
 import io.github.easyretrofit.extension.retry.core.interceptor.RetryExceptionFallBackHandler;
 import io.github.easyretrofit.extension.retry.core.interceptor.RetryInterceptor;
